version 4

+++ model.py
@@ -25,37 +25,80 @@ class Conexion():
             resp["mensaje"] = "No hay carreras registradas"
         return resp
 
+    def consultarCarreraSiglas(self, siglas):
+        resp = {"estatus": "", "mensaje": ""}
+        res = self.bd.vCarreras.find_one({"siglas": siglas})
+        if res:
+            resp["estatus"] = "OK"
+            resp["mensaje"] = "Carrera encontrada"
+            resp["carrera"] = res
+        else:
+            resp["estatus"] = "Error"
+            resp["mensaje"] = "No hay carreras registradas con esas siglas"
+        return resp
+
     def insertar_carrera(self, carrera):
         respuesta = {"Estatus": "", "Mensaje": ""}
-        carrera["estatus"] = "A"
-        self.col.insert_one(carrera)
-        respuesta["Estatus"] = "OK"
-        respuesta["Mensaje"] = "Carrera agregada correctamente"
+        administrador = self.bd.usuarios.find_one(
+            {"tipo": "A", "estatus": "A", "administrativo.estatus": "A"},
+            projection={"administrativo": True, "_id": False})
+        if administrador:
+            carrera["estatus"] = "A"
+            self.col.insert_one(carrera)
+            respuesta["Estatus"] = "OK"
+            respuesta["Mensaje"] = "Carrera agregada correctamente"
+        else:
+            respuesta["Estatus"] = "Error"
+            respuesta["Mensaje"] = "El usuario no existe"
         return respuesta
 
     def modificarCarrera(self, data):
         resp = {"estatus:": "", "mensaje:": ""}
-        res = self.bd.carreras.find_one({"_id": data["_id"]})
-        print(res)
-        if res:
-            self.bd.carreras.update_one({"_id": data["_id"]}, {"$set": data})
-            resp["estatus:"] = "OK"
-            resp["mensaje:"] = "Se actualizo la carrera correctamente"
+        estatus = self.bd.carreras.find_one(
+            {"_id": data["_id"]},
+            projection={"estatus": True})
+        print(estatus)
+        if estatus != None:
+            if estatus.get("estatus") == "A":
+                res = self.bd.carreras.find_one({"_id": data["_id"]})
+                print(res)
+                if res:
+                    self.bd.carreras.update_one({"_id": data["_id"]}, {"$set": data})
+                    print(data)
+                    resp["estatus:"] = "OK"
+                    resp["mensaje:"] = "Se actualizo la carrera correctamente"
+                else:
+                    resp["estatus:"] = "Error"
+                    resp["mensaje:"] = "La carrera esta Inactiva"
+            else:
+                resp["estatus:"] = "Error"
+                resp["mensaje:"] = "El estatus de la carrera no se encuentra Activa"
         else:
             resp["estatus:"] = "Error"
-            resp["mensaje:"] = "La carrera no existe"
+            resp["mensaje:"] = "El id de la carrera no existe"
         return resp
 
     def eliminarCarrera(self, id):
         resp = {"estatus": "", "mensaje": ""}
-        res = self.bd.carreras.delete_one({"_id": id})
-        print(res)
-        if res.deleted_count > 0:
-            resp["estatus"] = "OK"
-            resp["mensaje"] = "La carrera se elimino con exito"
+        estatus = self.bd.carreras.find_one(
+            {"_id": id},
+            projection={"estatus": True})
+        print(estatus)
+        if estatus != None:
+            if estatus.get("estatus") == "I":
+                res = self.bd.carreras.delete_one({"_id": id})
+                if res.deleted_count > 0:
+                    resp["estatus"] = "OK"
+                    resp["mensaje"] = "La carrera se elimino con exito"
+                else:
+                    resp["estatus"] = "Error"
+                    resp["mensaje"] = "No se pudo eliminar"
+            else:
+                resp["estatus:"] = "Error"
+                resp["mensaje:"] = "El estatus de la carrera no se encuentra Inactiva"
         else:
-            resp["estatus"] = "Error"
-            resp["mensaje"] = "La carrera no existe o no se encuentra en estatus de Inactiva"
+            resp["estatus:"] = "Error"
+            resp["mensaje:"] = "El id de la carrera no existe"
         return resp
